Renamed main => core in Pongo to align with introduced in Dumbo convention
Renamed postgresCollection to pongoCollection, as it'll be generalised in the follow-up commit
