+++ src/packages/pongo/src/core/typing/entries.ts
