+++ src/packages/pongo/src/index.ts
@@ -1,3 +1,3 @@
-export * from './main';
+export * from './core';
 export * from './mongo';
 export * from './postgres';
